Added a littlefs-fuse specific version number
Previously this was fixed to the underlying littlefs core version, but
that state of things doesn't work if you ever want new features in
littelfs-fuse outside of littlefs minor releases.

This currently matches the underlying core minor version, to tie things
together, but will fall out of sync as things are updated.